Sync `Instance::resolve` with the projection code
